fix(sov-core): CHERN_TARGET f64 死常数 → i32 整型, 与 C++ lcm_constants.h 对齐
C++ 侧: inline constexpr int CHERN_TARGET = 2 (整型)
Rust 侧原为 f64 2.0 且全库无引用 — 禁浮点精神遗留, 改为 i32 2
(Q16 版 131072 已在, 两轨约定镜像不变: sov_core +2 / sov_guard L7 −2)

+++ rust/sov-math/sov-core/src/constants.rs
@@ -32,7 +32,7 @@ pub const POLAR_WINDING: usize = 144;
 pub const TOROIDAL_WINDING: usize = 46;
 
 /// 陈数
-pub const CHERN_TARGET: f64 = 2.0;
+pub const CHERN_TARGET: i32 = 2;      // 与 C++ lcm_constants.h int CHERN_TARGET=2 对齐 (整型, 禁浮点)
 pub const CHERN_TARGET_Q16: i32 = 131072; // 2.0 × 65536
 pub const DELTA_Q16: i32 = 113506;         // √3 × 65536
